HW9 Files

+++ turtlebot-master/GetUserAngle.py
@@ -1,8 +1,8 @@
 import rospy
-from std_msgs.msg import String
+from std_msgs.msg import Int16
 
 def getAngle():
-    pub = rospy.Publisher('chatter', int, queue_size=10)
+    pub = rospy.Publisher('chatter', Int16, queue_size=10)
     rospy.init_node('talker', anonymous=True)
     #rate = rospy.Rate(10) # 10hz
     while not rospy.is_shutdown():
